feat: stable u32 UIDs, DAG-native snapshots, animated SVG timeline
Major rework of the trace visualization system:

Stable node identity:
- DagSnapshotNode.uid: u32 monotonic ID that persists across snapshots
- merged_from: Vec<u32> tracks coalescence (which UIDs were merged)
- added/removed flags for diff-style change tracking
- DagSnapshotEdge uses u32 UIDs instead of string names

DAG snapshot improvements:
- DagSnapshot gains reason: String explaining why this snapshot differs
- record_snapshot() now takes uid_start counter and reason
- record_dag_snapshot() for explicit edge topology with UIDs
- to_text() shows UIDs inline: "zenlayout.crop(#3)"

Animated SVG (to_animated_svg):
- CSS @keyframes animation, no JavaScript required
- Nodes slide to new positions across snapshots
- Nodes fade in/out when added/removed
- Color-coded by node kind (geometry=blue, filter=purple, etc.)
- Timeline bar at bottom with frame labels and animated marker
- Each frame holds for 2s, loops infinitely
- Works in browsers; static frames viewable in feh/eog

+++ src/graph.rs
@@ -956,8 +956,8 @@ impl PipelineGraph {
                 .edges
                 .iter()
                 .map(|e| crate::trace::DagSnapshotEdge {
-                    from: e.from,
-                    to: e.to,
+                    from: e.from as u32,
+                    to: e.to as u32,
                     kind: match e.kind {
                         EdgeKind::Input => alloc::string::String::from("input"),
                         EdgeKind::Canvas => alloc::string::String::from("canvas"),
